requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
sentence-transformers
faiss-cpu
networkx
lightgbm
joblib
openai
